refactor(utils,meta): move metaprogramming types to separate header
